+++ src/service/player.rs
@@ -11,7 +11,7 @@ pub async fn create_player_profile(conn: &MySqlPool, client: &Client, player: Pl
     let user_for_creation = PlayerForCreationDto::convert_player_into_user_for_creation(&player);
 
     let persisted_token = unwrap_or_return_handled_error!(create_user(client, &user_for_creation).await, Token);
-    let player_to_persist = Player::new_from_creation_dto(&player, &persisted_token.user_id);
+    let player_to_persist = Player::new_from_creation_dto(&player, &(persisted_token.user_id as u32));
     unwrap_or_return_handled_error!(500, player_dao::insert_player(conn, player_to_persist).await, Token);
     TypedHttpResponse::return_standard_response(200, persisted_token)
 }
@@ -21,7 +21,7 @@ pub async fn create_player_profile(conn: &MySqlPool, client: &Client, player: Pl
 pub async fn edit_player_profile(conn: &MySqlPool, client: &Client, player: PlayerForUpdateDto) -> TypedHttpResponse<Player> {
     let persisted_user = unwrap_or_return_handled_error!(user_service::authenticate_user_with_token(client, &UserForAuthenticationDto { app: APP_NAME.to_string(), id: player.user_id.to_string(), token: player.auth_token.clone() }).await, Player);
     //  Attempt to find player in database with the user id that user service gave back
-    let persisted_player = match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, persisted_user.id).await, Player) {
+    let persisted_player = match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, persisted_user.id as u32).await, Player) {
             Some(found_player) => found_player,
             None => return TypedHttpResponse::return_standard_error(404, MessageResource::new_from_str("Could not find player with id. Something went wrong.")),
         };
@@ -35,13 +35,13 @@ pub async fn login(conn: &MySqlPool, client: &Client, mut user: UserForLoginDto)
     user.app = APP_NAME.to_string();
     let persisted_token = unwrap_or_return_handled_error!(user_service::authenticate_user_with_password(client, &user).await, Token);
 
-    match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, persisted_token.user_id).await, Token) {
+    match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, persisted_token.user_id as u32).await, Token) {
         Some(_) => TypedHttpResponse::return_standard_response(200, persisted_token),
         None => TypedHttpResponse::return_standard_error(404, MessageResource::new_from_str("Could not find player with id. Something went wrong.")),
     }
 }
 
-pub async fn get_player_profile(conn: &MySqlPool, player_id: i32) -> TypedHttpResponse<PlayerProfileDto> {
+pub async fn get_player_profile(conn: &MySqlPool, player_id: u32) -> TypedHttpResponse<PlayerProfileDto> {
     let persisted_player = match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, player_id).await, PlayerProfileDto) {
         Some(player) => player,
         None => return TypedHttpResponse::return_standard_error(404, MessageResource::new_from_str("Could not find player with id. Something went wrong.")),
@@ -53,7 +53,7 @@ pub async fn get_player_profile(conn: &MySqlPool, player_id: i32) -> TypedHttpRe
     TypedHttpResponse::return_standard_response(200, PlayerProfileDto::new_from_player_and_counts(&persisted_player, trusted_player_count.count, trusted_by_player_count.count))
 }
 
-pub async fn get_player_trusted_list(conn: &MySqlPool, player_id: i32) -> TypedHttpResponse<Vec<Player>> {
+pub async fn get_player_trusted_list(conn: &MySqlPool, player_id: u32) -> TypedHttpResponse<Vec<Player>> {
     match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, player_id).await, Vec<Player>) {
         Some(player) => player,
         None => return TypedHttpResponse::return_standard_error(404, MessageResource::new_from_str("Could not find player with id. Something went wrong.")),
@@ -65,7 +65,7 @@ pub async fn get_player_trusted_list(conn: &MySqlPool, player_id: i32) -> TypedH
     }
 }
 
-pub async fn get_player_trusted_by_list(conn: &MySqlPool, player_id: i32) -> TypedHttpResponse<Vec<Player>> {
+pub async fn get_player_trusted_by_list(conn: &MySqlPool, player_id: u32) -> TypedHttpResponse<Vec<Player>> {
     match unwrap_or_return_handled_error!(player_dao::get_player_with_id(conn, player_id).await, Vec<Player>) {
         Some(player) => player,
         None => return TypedHttpResponse::return_standard_error(404, MessageResource::new_from_str("Could not find player with id. Something went wrong.")),
